Source: Make hash_key in tt entry 32 bits
bench: 4784662

+++ Source/structs.h
@@ -5,7 +5,7 @@
 #include <stdint.h>
 
 typedef struct tt_entry {
-  uint64_t hash_key; // "almost" unique chess position identifier
+  uint32_t hash_key; // "almost" unique chess position identifier
   int score;         // score (alpha/beta/PV)
   int move;
   uint8_t depth; // current search depth
